refactor(node-stdlib): Remove core module, move Result into global
