+++ packages/node-stdlib/src/index.ts
@@ -1,17 +1,7 @@
 // Copyright (c) 2020 Christopher Szatmary <cs@christopherszatmary.com>
 // All rights reserved. MIT License.
 
-// This rule is causing false positives so disable it in this file.
-// It's complaining that some modules export the same names.
-// This isn't an issue though because the exports are scoped
-// under the module name.
-/* eslint-disable import/export */
-
 export * from "./global";
-
-// core is a special case since it contains core types
-// that will likely be used frequently
-export * from "./core/mod";
 
 export * as bytes from "./bytes/mod";
 export * as cmd from "./cmd/mod";
